git-svn-id: http://svn.ilias.de/svn/ilias/trunk@5778 21b2c9ec-7c21-0410-8b45-9bfb8ed2bfc5

+++ survey/classes/class.SurveyQuestion.php
@@ -1004,7 +1004,7 @@ class SurveyQuestion {
 		$result = $this->ilias->db->query($query);
 
 		$directory = CLIENT_WEB_DIR . "/survey/" . $obj_id . "/$question_id";
-		if (is_dir($directory))
+		if (preg_match("/\d+/", $obj_id) and preg_match("/\d+/", $question_id) and is_dir($directory))
 		{
 			$directory = escapeshellarg($directory);
 			exec("rm -rf $directory");
